backend/requirements.txt
djangorestframework
Django~=3.1.3
djangorestframework_simplejwt
djoser
django-rest_auth
django-cors-headers
django-allauth